Added check_dir and download_jar functions
check_dir() runs from main() before menu(). It checks that the directory
$HOME/.mc_server_launcher exists and creates it if necessary. It then checks
if the server.jar file exists. If not, it calls download_jar.

download_jar(path) takes a Path object pointing to where the server.jar should
be saved. It grabs a json file from mojang that includes the latest
version and the download link associated.